feat: allow create new account by json import in onboarding
