+++ apps/extension/src/ui/apps/dashboard/index.tsx
@@ -19,7 +19,7 @@ import { Navigate, Route, Routes, useMatch } from "react-router-dom"
 import { DashboardLayout } from "./layout/DashboardLayout"
 import { AccountAddMenu } from "./routes/AccountAdd"
 import { AccountAddDerivedPage } from "./routes/AccountAdd/AccountAddDerivedPage"
-import { AccountAddJsonPage } from "./routes/AccountAdd/AccountAddJson"
+import { AccountAddJsonPage } from "./routes/AccountAdd/AccountAddJsonPage"
 import { AccountAddQrWizard } from "./routes/AccountAdd/AccountAddQrWizard"
 import { AccountAddSecretWizard } from "./routes/AccountAdd/AccountAddSecretWizard"
 import { AccountAddWatchedPage } from "./routes/AccountAdd/AccountAddWatchedPage"
