+++ apps/extension/src/ui/apps/dashboard/routes/AccountAdd/AccountAddJsonPage.tsx
@@ -1,0 +1,23 @@
+import { HeaderBlock } from "@talisman/components/HeaderBlock"
+import { Spacer } from "@talisman/components/Spacer"
+import { AccountAddJson } from "@ui/domains/Account/AccountCreate/AccountAddJson"
+import { useSelectAccountAndNavigate } from "@ui/hooks/useSelectAccountAndNavigate"
+import { useTranslation } from "react-i18next"
+
+import { DashboardLayout } from "../../layout/DashboardLayout"
+
+export const AccountAddJsonPage = () => {
+  const { t } = useTranslation("admin")
+  const { setAddress } = useSelectAccountAndNavigate("/portfolio")
+
+  return (
+    <DashboardLayout withBack centered>
+      <HeaderBlock
+        title={t("Import JSON")}
+        text={t("Please choose the .json file you exported from Polkadot.js or Talisman")}
+      />
+      <Spacer />
+      <AccountAddJson onSuccess={setAddress} />
+    </DashboardLayout>
+  )
+}
